Add "enableScreenshotCapture" option to CordProvider
This is to allow customers to disable screenshotting capturing. I am
naming this "enableScreenshotCapture" rather than
"enableScreenshots", because the latter might imply that screenshots in
messages of other users would not be visible either.


Test Plan:
On testbed I created an annotation with sidebar. A screenshot was not
taken when enableScreenshotCapture=false

Reviewers: flooey, mhtien, yeo-yeo

Reviewed By: yeo-yeo

Pull Request: https://github.com/getcord/monorepo/pull/3931

monorepo-commit: a64b069107880abd7deb2bea35edccc162014f95