Add onLoading and onRender to ThreadList
Same as the callbacks on Thread. Clients can use these to render a
different component while the ThreadList is loading.

Just like for thread we had to re-dispatch both loading and render
events when the threadID changed, for threadlist we need to re-dispatch
both events when the location changes since we are fetching a new set of
threads.


Reviewers: AlbertQM, yeo-yeo, lazybean, flooey, jwatzman

Reviewed By: jwatzman

Pull Request: https://github.com/getcord/monorepo/pull/3962

monorepo-commit: 4e56a4ce193d224a2917febfa99851d70a3c134d