Adding a showExpanded prop to the composer component
Whilst experimenting with the composer component, there were cases where I wanted to have the composer expanded by default in the modality I was building. I think this is a reasonable enough ask for any developer to have, thus I have exposed this prop in the composer component.

I thought the name `showExpanded`, which is used by us internally as well was not bad, but please do let me know if you think this prop should be named differently.

Test Plan:
In the testbed composers, I set this prop to true and the composer was always in its expanded state.


Reviewers: jwatzman, flooey, lazybean

Reviewed By: lazybean

Pull Request: https://github.com/getcord/monorepo/pull/4000

monorepo-commit: 3e9e9cdf882377276b9260a9a6b0e7d3561e0e38