@cord-sdk/types: remove dependency on lodash
Depending on lodash is causing a bit of a pain. Also it is a big package to depend on, and we only need it in one place to check equality of Locations.
This commit adds a function "isEqualLocation" to our types package and uses that instead of lodash's isEqual.
It makes use of the fact that we know that the Locations are key/value objects where the values are flat values (string, number or boolean).

In order to do this properly, this commit also adds a unit test to the types package. It introduces "jest" as a dev dependency for tests.


Reviewers: flooey

Reviewed By: flooey

Pull Request: https://github.com/getcord/monorepo/pull/4032

monorepo-commit: 3a206da5d3b7d2e1100979b087983497b7b48beb