Avoid default imports when package defines no default
With TypeScript's `esModuleInterop` setting set to true (as in our config), TypeScript actually correct for this automatically, but in some contexts or with some build tools it might be a problem that we don't import some packages correctly. This commit fixes that.


Reviewers: flooey, jwatzman

Reviewed By: jwatzman

Pull Request: https://github.com/getcord/monorepo/pull/4031

monorepo-commit: 82c853721196c73e967f799b6d6b5bf4382cd6ea