Add observePresence function (#4030)
The presence API was our first JS API and didn't follow our
later-established convention of providing an observeFoo() function to
get live updating values, instead supplying separate getPresent() and
addListener() functions and relying on the caller to reconcile those
values.  Add an observePresence() function to follow the pattern.


Also reimplement our web components other than MultipleCursors in
terms of observePresence(), which makes the PresenceReducer code
significantly simpler, which is good confirmation this is a better
pattern.

Test Plan:

Tested PagePresence on the testbed with a different user in two tabs.
Confirmed that it worked properly normally, continued to exclude the
viewer when that option is set, and worked with a subpage when partial
matching was enabled.

monorepo-commit: 3a1dff0ff4f6e42fbf15afe57bb0b77b2b5129d7