Comments: Introducing a new thread mode for comments
We want to create threads that look more like comments. Something like this:
<img width="278" alt="image" src="https://user-images.githubusercontent.com/102963329/226970244-a866c0a0-561e-4a04-98f3-f230a91bbc93.png">

To do that, we need to modify our existing threads to accommodate for these changes. This PR:
1. Creates a new thread mode, called: `comment` and a respective `<CommentThread />` component to go with it
2. Implements this component as part of the thread list to test it out in the relevant context
3. Creates a feature flag in the ThreadList, which is only enabled for us internally, to toggle between how it presents threads

Note: The reason I decided to expose a prop, which is unusable unless you are in your local env, is so that we can experience using the thread list as a real end user would and hopefully find any flaws. I haven't documented this prop anywhere and I am not planning to, until we decide it's the right one.

Q: Why create a new thread mode instead of modifying the InlineThread somehow?
A: I have copied a bunch of code from the InlineThread component, which makes this a fair question. Even though from a UI perspective, the CommentThread and InlineThread look very similar, their implementations have significant differences. For example, the CommentThread completely ignores the concept of message blocks (that is, if a user sends a bunch of consecutive messages, we bundle them). The CommentThread will also have an avatar together with the composer, so that it is clear who is replying and at which thread - this was specifically created for the case where we have many CommentThreads stacked.

Q: So if we decide to make changes to the thread UI (css), we will need to almost make identical changes to the CommentThread and InlineThread
A: Even though right now I could have simply added an insignificant number of conditionals in the InlineThread to get to where I needed to be, my gut tells me that these two components will diverge drastically in terms of inner working, rather than UI. So I would push that we rather did the same UI change in two components, instead of adding conditionals within the InlineThread to make a change, which will slowly lead us down the spaghetti code road.

Q: New ThreadList mode prop. What? why?
A: Our goal is to eventually use the CommentThread in a `<Comments />` component, where it makes sense to use the ThreadList. Currently ThreadList only allows collapsed threads. This prop takes and enum which is either `collapsed` or `comments`. I considered adding a boolean called something like: `allowReplies`. However, I thought that in the future we might want to add another mode. For example, the ability to expand threads and read them, but not be able to reply. So I opted for this enum. I am open to any suggestions though 🙂

Test Plan:
(Images attached)
![image](https://user-images.githubusercontent.com/102963329/226975913-e095efbd-fa89-4f1a-aac1-c699613e741e.png)
![image](https://user-images.githubusercontent.com/102963329/226975943-5e4844de-afa1-4ab5-b44d-8b65c49f3f51.png)



Reviewers: Netceer, flooey, jwatzman, AlbertQM

Reviewed By: jwatzman

Pull Request: https://github.com/getcord/monorepo/pull/4035

monorepo-commit: 529366a774af1772151736ad6a3e6296b416622f