ParticipantFacepile: Creating experimental component
We will need this as a building block to create a thread collapsed state, but I see other uses
for sticking a facepile with the participants of a thread. Maybe in a thread preview, together
with badging? Maybe in a thread launcher type button?

In any case, I will be playing with this one for the next few days.

I can also see the usecase where we need thread participants per location. Our immediate
use case is thread-ids, so I will postpone location. But if this proves useful, we should
definitely allow location as well.

Test Plan:
Added in testbed - played around with the props - works as expected.


Reviewers: flooey, jwatzman

Reviewed By: jwatzman

Pull Request: https://github.com/getcord/monorepo/pull/4105

monorepo-commit: d5b671987f6d59f8b6d421f886874ba6b891e359