Rename presence.observePresence to presence.observeLocationData
One of several upcoming renames to make our API more consistent. This
does the rename with backwards-compat shims in place so that existing
code continues to work, and so that the not-yet-updated docs remain
technically correct. Once all of the renames have landed, the plan is to
update the docs (hopefully with a slightly larger rework of the relevant
sections) and release a new npm package.


Reviewers: zopu, lazybean, flooey

Reviewed By: flooey

Pull Request: https://github.com/getcord/monorepo/pull/4201

monorepo-commit: 634fbcbb455a9b2414bc90337c19946062d4e260