Rename notification.observeNotificationSummary to notification.observeSummary
One of several upcoming renames to make our API more consistent. This
does the rename with backwards-compat shims in place so that existing
code continues to work, and so that the not-yet-updated docs remain
technically correct. Once all of the renames have landed, the plan is to
update the docs (hopefully with a slightly larger rework of the relevant
sections) and release a new npm package.


Reviewers: zopu, flooey, lazybean

Reviewed By: flooey

Pull Request: https://github.com/getcord/monorepo/pull/4202

monorepo-commit: 91f451721f73a26d366b5c2899b558e5acdb93d9