Rename useCordThreadSummary to thread.useThreadSummary
One of several upcoming renames to make our API more consistent. This
does the rename with backwards-compat shims in place so that existing
code continues to work, and so that the not-yet-updated docs remain
technically correct. Once all of the renames have landed, the plan is to
update the docs (hopefully with a slightly larger rework of the relevant
sections) and release a new npm package.


Reviewers: lazybean, flooey, #engage

Reviewed By: lazybean

Pull Request: https://github.com/getcord/monorepo/pull/4205

monorepo-commit: 46b0cfe6eb8b6b04c6a0bcad6305eba5094a1929