Rename activity.observeThreadSummary to thread.observeLocationSummary
One of several upcoming renames to make our API more consistent. This
does the rename with backwards-compat shims in place so that existing
code continues to work, and so that the not-yet-updated docs remain
technically correct. Once all of the renames have landed, the plan is to
update the docs (hopefully with a slightly larger rework of the relevant
sections) and release a new npm package.


Reviewers: AlbertQM, lazybean, flooey

Reviewed By: AlbertQM

Pull Request: https://github.com/getcord/monorepo/pull/4206

monorepo-commit: 2f159da9bbf39482dcf254a97a7bcf2056f5a7db