Repurposing ThreadFacepile to Facepile component
The Facepile component takes an array of external user IDs as an input and returns a Facepile of these users, ordered in the way the IDs were provided in the array.

Note: We may want to provide an integration guide on how to build this component using the Avatar component and the presence hooks. But I am exposing this for now as it is more urgent.

Test Plan:
Added it to the testbed using the `useThreadSummary` hook and it correctly displayed a facepile with the users I provided.

![image](https://user-images.githubusercontent.com/102963329/230410860-a7ed3f0f-9158-4d85-9bd8-8c8495ff4926.png)



Reviewers: flooey, jwatzman

Reviewed By: jwatzman

Pull Request: https://github.com/getcord/monorepo/pull/4251

monorepo-commit: 1fba5896c33cf8b0a857d94eb63f74d8decb0632