Rename experimental.messages.observeThreadData to thread.observeThreadData
One of several upcoming renames to make our API more consistent. This
API was under experimental and undocumented, and so does not need any
backwards-compat shims. It is now out of experimental since we're
relatively happy with the API; it will be documented once all of the
renames have landed (hopefully with a slightly larger rework of the
relevant sections).


Reviewers: nickfil22, flooey, lazybean

Reviewed By: nickfil22

Pull Request: https://github.com/getcord/monorepo/pull/4270

monorepo-commit: a99f5c0e83b31f2b1734e825cd08d534cea647b7