Exposing the ThreadFacepile component
One potential application of the new `Facepile` component is to use it in combination with the thread data hook to display all the participants in a thread. This PR exposes this implementation in our SDK.

Test Plan:
Added it in the testbed, can see the correct faces.


Reviewers: jwatzman, flooey

Reviewed By: jwatzman

Pull Request: https://github.com/getcord/monorepo/pull/4281

monorepo-commit: 8e45a9e15738f31d66c346c8f87fad605044fb30