Presence: don't let ephemeral or ephemeral.locations be missing
I was updaing the docs for this page, and realised I didn't know what
the difference was between ephemeral being empty, vs the location list
being null or empty, in `observePresence`.

AFAICT there isn't one. (For the deprecated `addListener` which only
gives delta updates, having one missing mean "no update" makes sense,
but not for a function which always gives full data.) So let's make this
type more restrictive so that it's less confusing and people can rely on
this fact.

This change should be safe backwards-compat-wise both because this is a
new undocumented function (found this when writing the docs!) and
because code would already need to deal with empty lists etc anyway.


Reviewers: flooey

Reviewed By: flooey

Pull Request: https://github.com/getcord/monorepo/pull/4309

monorepo-commit: c85fbf8ba27c793c5f9d845b57d377c664a8152b