Add new highlightThreadId prop to thread list
Exposing this prop to allow more control from the dev user which thread
to highlight in the thread list. Devs just need to pass in a threadId
and we prioritise highlighting this threadId over any other highlight
defaults we may have with any other Cord component (e.g when a dev uses
the highlightOpenFloatingThreadID prop as well we will highlight the
threadId in highlightThreadId).


Test Plan:
Added an external thread ID in the input field in the test bed and if it
exists then it is highlighted. If not then it is not. If left blank
and you click on the floating thread then this highlights.

https://user-images.githubusercontent.com/59407030/229448381-af8858ca-2fc7-4c4f-ba01-cff4d025f110.mov

Reviewers: flooey, yeo-yeo

Reviewed By: yeo-yeo

Pull Request: https://github.com/getcord/monorepo/pull/4203

monorepo-commit: f977e973b04396c48cfe0657fa4437ec5e06e8ba