Removing file which was not correctly deleted in previous commit
I had initially created the `ThreadFacepile` component within a different directory called:
`compositeComponents`. When deciding to actually move this inside `components`, in #4281, I drag/dropped
it from the VSCode UI, which for some reason was detected as a 'cp', not an 'mv'. This PR fixes this mistake
by deleting the unnecessary file.

Test Plan:
No import complained that the ThreadFacepile did not exist inside `compositeComponents`. Also checked
the ThreadFacepile component, which was correctly imported from `components` in both places it needed to be
imported in.


Reviewers: jwatzman, flooey

Reviewed By: jwatzman

Pull Request: https://github.com/getcord/monorepo/pull/4316

monorepo-commit: 4dbff4bdbd370adcb44e0ef8cf19161bfaa459cc