Adding filtering/sorting options to the observeThreadIDs API
Adding two options to the observeThreadIDs API, the option to sort via the timestamp of the first message of a thread and the option to filter messages based on whether they are resolved or not.

Test Plan:
In the testbed - I played around with the two parameters and correctly got back the results I was expecting.


Reviewers: jwatzman

Reviewed By: jwatzman

Pull Request: https://github.com/getcord/monorepo/pull/4288

monorepo-commit: 2ab9d45aba84051ca5feefd8c76a5a2bc17461d8