Implement API for customizing MultipleCursors positions (#4020)
Add an API where developers can customize the locations sent by
MultipleCursors, which makes it useable in many more situations.  In
particular, in apps with internal coordinate systems (like canvas
apps), this will let them easily show cursors in the right place.

The API consists of a single setter which sets a pair of functions.
The first function maps mouse events to a Location; by default, this
is just the page X and Y.  The second function maps Locations
transmitted through the presence system back to document coordinates
on the receiving system.  These are set at the same time to reinforce
that they need to match; if the default implementation is used for one
half but not the other, likely nothing will work.

monorepo-commit: 7bd0fa9efa7343fbd78c829245cc5ca416c51d5c