Adding tooltip prop to the Facepile component
Note: I recommend `hide whitespace` in the diff

As discussed in the launch review, it would be good to have a tooltip
that can be enabled/disabled in both the Avatar and Facepile.

Note: I realise that the Facepile2 component is a bit of spaghetti, but I really
don't want to spend any more time on this, especially given that we will probably
scrap Facepile2 in the next few weeks for a leaner component, with vanilla css
as well.

Test Plan:
Testbed Facepile behaved correctly when enabling/disabling the prop


Reviewers: flooey, jwatzman

Reviewed By: jwatzman

Pull Request: https://github.com/getcord/monorepo/pull/4502

monorepo-commit: dbd6a69d05b1c288b40eeedc1ab2bc8e0c4bd3f6