Unify thread REST API and JS API
Let's make sure that the REST API and the JS API for thread summary have
the same types and so return the same things as much as possible. There
are a few bits of data specific to each one, e.g., the JS API has
viewer-specific data which doesn't make sense in the REST API.

Add in `name` while I'm here because it's easy and we somehow forgot it
in all the previous iterations on this we've been doing!


Reviewers: nickfil22, flooey

Reviewed By: flooey

Pull Request: https://github.com/getcord/monorepo/pull/4508

monorepo-commit: 1a5e20ee0808fd3f9f619aa5ebd35ef63d7b1532