Fix error with null names (#4544)
Some users in our DB have null names, which we didn't expect, and
causes errors at the GraphQL schema level.  Fix the schema as a stop
gap, then we can fix this problem for real.

Test Plan:

Set a user's name to null in my local DB and opened testbed on master,
saw the error.  Ran again with this fix, no error.

monorepo-commit: 38a5f3193d284ce4c76ffe64be2e49eee42651d7