REST API: PUT endpiont /v1/threads/:threadID
Creates an endpoint which can update a thread. The actual code backing
that is pretty straightforward, but there are a number of changes needed
to make the ajv validation work out.

Because we're sharing types with GraphQL (yay!) this means we now have a
dependency between our `types` and `api-types` package -- the latter
depends on the former (which seemed the right way to do it). So
`package.json` is updated to reflect this.

This in turn requires a couple of updates to `generate.mjs` to deal with
this dependency as well as other TS syntax that appears in `types` which
the script previously didn't need to deal with. (This is nothing
interesting, `generate.mjs` wasn't set up properly, it just never came
up before.)

Related, I had to change the definition of `FlatJsonObject` to an
equivalent one using object syntax instead of `Record` so that ajv
generates the correct schema. (*This* is probably a bug in ajv, but
again which never came up before, since this is the first appearance of
a Location as an input in our REST API.)


Reviewers: jozef-mokry, yeo-yeo

Reviewed By: jozef-mokry

Pull Request: https://github.com/getcord/monorepo/pull/4551

monorepo-commit: e1990150994861b4d05138863a44827245fdf8e8