Extract TSDoc comments for use in docs (#4597)
I have always disliked that what should be doc comments on our public
interfaces end up only in the docs site code, so make it so we can
write them as TSDoc comments and then use those comments in the docs
site instead.  Involves using api-extractor to parse and extract all
the TSDoc comments, then a script of ours to turn those into the same
data format that we already use for AJV-generated types for the REST
API.  Then that can be imported directly in our docs pages.

To show it works, this diff includes a proof-of-concept of documenting
the thread.observeLocationSummary JS function and using that in the
code.

One side effect of doing this is that we need to change types that we
want to document from using `type` to using `interface`.  For arcane
reasons I don't entirely understand, types come through into the
extracted API data just as opaque values (I guess because types can be
arbtrarily complicated?), so if we want to document types, we need
them to be interfaces.  That's fine, types and interfaces are largely
interchangeable, it's just a little annoying.

Test Plan:

Viewed the docs, there are some minor formatting changes and at least
one bug fix, but it mostly looks the same.

monorepo-commit: 0f45ce49cc22b4cd0883d2bdde716e8f5490a2b3