Remove "type" input from CreateThreadMessageHandler
We shouldn't allow writing this field -- it lets you write a message
saying "X resolved this thread" but doesn't actually set the thread as
resolved.

We could make this resolve the thread if written, but even then, this
lets you write a thread resolved message with nonempty content, which
again is very unlikely to do what you expect (the content will be
ignored).

Let's keep this as a read-only quasi-implementation-detail for now since
it's so likely to be a footgun.


Reviewers: jacksongabbard, flooey

Reviewed By: flooey

Pull Request: https://github.com/getcord/monorepo/pull/4686

monorepo-commit: 145fa1bad7fe28d89b41f738d24fcf1e55e2ed3f