REST API: PUT /v1/threads/:threadID inject resolved/unresolved messages
When resolving/unresolving a thread, we should not just set the thread
status, but also inject a message into the thread indicating to the
reader what happened -- and *who* resolved/unresolved the thread.
Unfortunately the "who" requires a userID, which the caller will have to
specify. That argument is fully optional; if not provided, no such
message will be injected.


Reviewers: flooey

Reviewed By: flooey

Pull Request: https://github.com/getcord/monorepo/pull/4711

monorepo-commit: 372c9fe19426dcb3d802b092b3c24bbe3d40d66d