Aplication API: delete application endpoint
This was originally on #4611 but I've moved it here as we’ve changed the
way we want to endpoint to function.

Using this endpoint will result in a complete deletion of the app and
all the connected entities. We have a required parameter for the app's
'secret' to minimise any chances of accidental deletions.


```

Reviewers: flooey, dmmiller

Reviewed By: dmmiller

Pull Request: https://github.com/getcord/monorepo/pull/4656

monorepo-commit: b6e296e8798648fb92430adb9aa7e4175d88d09e