Updating Typescript in our sdk-js package
Until now, we used to rely on `rollup` to include typescript, but #4710 will need
to make use of the `satisfies` operator which is included after ts 4.9 (rollup locked to 4.8.4).

Needed to update `typescript-json-schema` as they had a bug that was revealed with
the ts upgrade.

I also added the vanilla extract dependency which will be needed for writing css for our oss
components.

Test Plan:
Everything builds


Reviewers: jwatzman, flooey

Reviewed By: jwatzman, flooey

Pull Request: https://github.com/getcord/monorepo/pull/4726

monorepo-commit: e6f910459b229372a211cd2ecf5b2606e77bd3af