Allow cord-pin to have children
It is expected that developers will want to show some component next to
the cord-pin, usually cord-thread.

This PR makes it possible for devs to write:
```
<Pin>
  <Thread />
</Pin>
```
or
```
<cord-pin>
  <cord-thread></cord-thread>
</cord-pin>
```

By default Pin will position the children absolutely below the pin, but
developers can easily overwrite this styling either with inline styles
or with a CSS selector in THEIR stylesheet. For example, if they want a
bigger gap between the pin and the thread:
```
cord-pin > cord-thread {
  top: 120%;
}
```


Reviewers: flooey, dmmiller, jwatzman, lazybean

Reviewed By: jwatzman

Pull Request: https://github.com/getcord/monorepo/pull/4738

monorepo-commit: ba6aceca86760b554d4723e4bee3e14068c2c718