Add support for action_message message type (#4803)
Sometimes things happen.  While I think we can all agree that this is
not a desirable state of affairs, we should still enable people to
share that information with others.

To that end, add a new kind of message type, `action_message`, which
renders a simplified display with no author header, no message menu,
etc.  This is intended to replace the `thread_resolved` and
`thread_unresolved` message types, which can be expressed in this
form, while also enabling our customers to include their own action
messages.  The `thread_resolved` and `thread_unresolved` types will
still live until we can run a migration to convert them all to
`action_message` messages (which should be mechanical, but is not
included here).

As part of this, add a new property to messages, `iconURL`, which lets
you set an icon for a message.  This is only used for action messages.
Add type and iconURL parameters to creating messages and fetching
them.

The interesting part of the change is in `ActionMessage2.tsx`, which
previously only rendered the `thread_resolved` and `thread_unresolved`
messages.  Now it can take an `action_message` message and will render
it through `renderNode`, which renders message content.  `renderNode`
takes an additional argument for what style of rendering we're using
(`normal` or `action_message`), with the latter not applying the bold
style or @ prefix for user mentions to keep the same formatting as the
existing display of `thread_resolved`, etc.

Test Plan:

<img width="391" alt="Screenshot 2023-05-30 at 9 42 53 AM" src="https://github.com/getcord/monorepo/assets/215991/44021429-c99a-493c-9939-f968069a6ce8">

monorepo-commit: 9c2194e79b977ce7658bd2a9abe45bc76532b5a4