Fixing typos in fonts.css.ts
I made a typo when converting the styles to globalStyles so the fonts didn't
target the right classes. This mistake had no impact, as these fonts aren't used anywhere.
That was also probably why it was not caught.

Test Plan:
Tested in #4696


Reviewers: AlbertQM, flooey, #sdk

Reviewed By: AlbertQM

Pull Request: https://github.com/getcord/monorepo/pull/4824

monorepo-commit: f247113e89495a1a9d71e7001bd0abfdfcc682d3