Comments 4: Adding a composer to reply in a thread
Moving the existing `ThreadedComments` component to our opensource package and renaming it to `Comments`.

This PR completes the functionality of replies. It renders a composer with an avatar when clicking on the reply button. The close button on the composer will make the composer disappear and the `Reply` button appear again.

An odd thing I've done is that I have duplicated the Reply button in the CollapsedReplies and ThreadReplies components. The reason I've done this is because adding it to the top-level thread component and sharing it between the two leads to some gnarly logic to handle all states, which is super confusing (If you are curious, check out the `ThreadedComments.tsx` component - I did that over there). I thought a duplication made things way more readable and was hence worth it.

Note: Autoscrolling and focusing the composer is when opening a thread is pending. I will tackle this in a next PR.

Test Plan:
https://github.com/getcord/monorepo/assets/102963329/860c8663-3662-4f00-b84a-25898d2a0a85





Reviewers: flooey, jwatzman, lazybean, AlbertQM

Reviewed By: jwatzman

Pull Request: https://github.com/getcord/monorepo/pull/4815

monorepo-commit: 2cfdc4fb0aabf8c0ffc68d18322188237e9c4a06