RFC: custom renderers for message nodes
Now that we are exposing our message format via the REST API, it might
be useful for customers to be able to inject custom items into our
messages. (This came up the other day in the context of folks using Cord
for AI chat bots.)

We've spitballed how to do this more than once, so I sat down and pulled
this together, since I don't think it's that difficult or controversial.
The idea is that we'll be provided a dictionary of custom renderers,
where the key is the node type it's for, and the value is a function
which takes that node and returns an `HTMLElement` for us to inject.

This is mostly relatively straightforward. There are a few annoying bits
though. For one, this first attempt only allows specifying the custom
renderers as a global array to `CordSDK.init`/`CordProvider`. The issue
is you can only send strings through to web components, so there's not a
good way of sending through these functions. There are some *ugly* ways
though, so if this global-ness becomes a problem, we can do that --
internally, our React contexts are already very well set up for partial
overrides so we aren't backing ourselves into a corner here.

Internally, since we're using React, we have to inject their
`HTMLElement` into our React DOM. There is, shockingly, not a great way
to do this. We need to have *some* element we `appendChild` the
`HTMLElement` to inside a `useEffect`, which adds an unfortunate and
useless extra `span` DOM node, but it is what it is. (Note that we are
using an `HTMLElement` here as the injected value, not a `string`, so
that -- among other reasons -- we don't lose event handlers and the
like. Not that using a `string` and the "unsafe set html" functionality
would help, it has the same problem with needing a parent node!)

Finally, if the customer is using React on their end, they'll need to
`ReactDOM.render` (or the React 18 equivalent I guess) in order to set
up React on an `HTMLElement` to send to us. This is not difficult *if*
you're familiar with how you actually set up React at the toplevel, but
is maybe arcane for folks who aren't. We can provide a helper for this
if we need -- but tbh I think there are actually better ways to do this,
which are even more arcane and are beyond me, so I haven't done that yet
so that we don't provide a crappy helper when something better is
available that really *does* do deep arcane magic.


Reviewers: flooey, yeo-yeo, jacksongabbard, lazybean, jozef-mokry

Reviewed By: jozef-mokry, flooey

Pull Request: https://github.com/getcord/monorepo/pull/4800

monorepo-commit: 06d85c8843a1c998e5b3402151beee125e90c845