Convert absolute links in TSDoc to relative (#4884)
Convert all the absolute markdown links in our TSDoc to relative links
at extraction time, so they can be relative in our docs (good for
local development) but absolute in the TSDoc (good for viewing in a
code editor)..

Test Plan:

Looked at
https://local.cord.com:8191/js-apis-and-hooks/presence-api/getPresent
(which uses the JS API data) and
https://local.cord.com:8191/reference/rest-api/batch (which uses the
REST API JSON schema).  Links in the properties lists all went to the
local host.

monorepo-commit: 1fcf25c720efa3eee124edce452f94e1ff223ba9