Notifications: GET /v1/users/:userID/notifications
A new endpoint, types, and documentation to list the notifications for a
specific user.

This is slightly inconsistent with the existing notifications endpoint,
which lives at `/v1/notifications` -- due to technical limitations, it's
simpler to only fetch notifications for a single particular user right
now, so I put this under `/v1/users`. This still leaves room for a
future `GET /v1/notifications` to get all notifs if we want to implement
that, as well as `POST /v1/users/:userID/notifications` to send to a
particular user. (But we'll also likely add a *list* of users to send
to, so keeping `POST /v1/notifications` as it exists today is also
useful.)

The layout of the types in the `api-types` package is slightly
idiosyncratic, and is done that way so that our automatic docs generator
generates useful output without tons of nesting, and also to appease the
fact that it wants everything to be exported. Unfortunate but not worth
investing in right now IMO.


Reviewers: flooey

Reviewed By: flooey

Pull Request: https://github.com/getcord/monorepo/pull/4918

monorepo-commit: 51cd309f7cd1f8ba0a8810ba3591e26e8de9b89f