Fix up message type definitions (#4945)
Adjust the message type definitions so they match what we're actually
returning (adding `organizationID` and `threadID`) and convert them
into interfaces so they get their propertyOrder emitted properly (no
idea why this is necessary, but it appears to be).

Also add type annotations into the code to enforce that they are
returning the correct types, and update the code slightly so it
matches the type.

Test Plan:

Everything compiles

monorepo-commit: d2335ae9f04d5e0c43feb54f8c2a44ced3b9035c