Stop using jsx-runtime in opensource/sdk-js (#4994)
In order to increase compatibility with potential customer build
configurations, stop using the new JSX transform for React[1].

There are two issues with the new transform.  One is that it's only
supported on React 17+, as well as some backported versions of 15 and
16.  The other is that, pre-React 18, the new transform involves an
import from `react/jsx-runtime`, but that file isn't explicitly
exported by the `react` package, and as a result some build systems
(notably, Webpack 5+) will not be able to process the import [2].

[1] https://legacy.reactjs.org/blog/2020/09/22/introducing-the-new-jsx-transform.html

[2] https://github.com/facebook/react/issues/20235#issuecomment-734808524


Test Plan:

npm run build

npm run tsc-packages

Ran the testbed, all components seemed to render fine, including
ThreadedComments.

monorepo-commit: 0c6f26eb5e8bc8bc3a4dc8d6a16325081d3ec473