cord-message: Remove shadowRoot
We're already released this component, telling devs they can customise it using CSS but it's not yet true.
To do so, we need to get rid of the shadowRoot.

Note: This component will have a mix of stable `classNames` and randomly generated ones.
This is because we haven't fully migrated all the components that are used to build the Message.

I've replaced WithTooltip with the new one while I was here..

Removing the `shadowRoot` means `:host` styles won't work anymore, so I've updated that.

Test Plan:
Load testbed, check Message looks OK.


Reviewers: lazybean, flooey

Reviewed By: lazybean

Pull Request: https://github.com/getcord/monorepo/pull/4985

monorepo-commit: ba4cfe0747ef8890356224af91d22074e9cb7769