ThreadedComments CSS: fix avatar targeting
We don't need a classname for the avatar, we can just target it.
Unfortunately we do need one for the wrapper of the composer/avatar
pair, which is a reasonable unit to have, but doesn't lend itself to a
great name, so leaving that alone.


Test Plan: Component looks the same as before.

Reviewers: AlbertQM, lazybean, flooey, nickfil22

Reviewed By: lazybean

Pull Request: https://github.com/getcord/monorepo/pull/5028

monorepo-commit: d6ab8475163c21bf93fa24a12d02c11d7ac33186