Add REST API for getting one application (#5072)
We supported getting all applications but not a specific one, might as
well add that.

Also in playing with the types, discovered that we allow setting the
iconURL but don't return it in the getters, which is a little weird,
so add that.

Test Plan:

API method: called, it worked


Docs: Mark I Eyeball

monorepo-commit: f6519acd80356da17aa248397d9fa0d8539ee7e1