Add typing indicators to REST API (#5096)
This data is already available in the client JS SDK, so add it to the
REST API as well.

This required a little bit of gymnastics in the list threads endpoint
to show with a minimal number of queries, but it only requires one
call to the DB to load the typing users (which will usually be a
no-op, since nobody will be typing).

Test Plan:

monorepo-commit: ab242961efa9e3985e3e7fa5dacc68a7c0eff401