Make the default max composer height smaller (#5144)
Our existing default max composer height doesn't work well for
situations like a thread in FloatingThreads or the AI chatbot where
the composer isn't in a full-height context; `40vh` might well be more
than the entire height of the thread.

Instead, limit it to `40vh` or `10em`, whichever is smaller.  What I
really want is 6-7 lines, but for that we'd need to use the `lh` unit
to specify a specific number of line heights, but that's not yet
supported in Firefox.  With our default 14px font-size and 20px
line-height, 10em is 7 lines.  This can still be overridden by
customers using the `--cord-composer-height-max` variables (or,
eventually, regular CSS).


Test Plan:

Wrote a comment with lots of newlines in a thread in in the testbed,
it expanded until about 5 lines were visible.

monorepo-commit: d2b3be1e01bdcf0969c2a8cc207bb39897595344