cord-notification-list-launcher: Remove `shadowRoot`
WebComponent has immediate effect (it's served from `latest.js`, so we default the `useShadowRoot` to `false`.

The React wrapper in the SDK is opt-in (developers need to update the SDK to get it), but by default they won't get the `shadowRoot` -- They can opt-out with the `useShadowRoot` property.

Test Plan:
Load `NotificationListLauncher` in the docs locally, check it looks the same as prod.


Reviewers: lazybean, flooey

Reviewed By: lazybean

Pull Request: https://github.com/getcord/monorepo/pull/5184

monorepo-commit: 09405b119f1ae651daaf13a42c428ba2f69e97be