cord-inbox-launcher: Remove `shadowRoot`
WebComponent has immediate effect (it's served from `latest.js`, so we default the `useShadowRoot` to `false`.

The React wrapper in the SDK is opt-in (developers need to update the SDK to get it), but by default they won't get the `shadowRoot` -- They can opt-out with the `useShadowRoot` property.

Test Plan:
Load `InboxLauncher` in the docs locally, check it looks the same as prod.


Reviewers: flooey, lazybean

Reviewed By: lazybean

Pull Request: https://github.com/getcord/monorepo/pull/5185

monorepo-commit: 741132aec73062693ea17050f1de81901021484d