Rewrite docs-extract-tsdoc.ts to use TS compiler (#5326)
Rewrite the docs-extract-tsdoc.ts script to use the TypeScript
compiler API to compile and extract the information about our various
types rather than the api-extractor package.

The big benefit here is to be able to extract arbitrarily-complex data
about our types.  In particular, it means we get nested object
information, whereas before we were limited to just the top-level
properties of a type, and also we can easily "see through" type
definitions like `Omit<Foo, 'bar'>` and see what the resulting set of
properties really is.  As our types become more complicated, this is
necessary to extract a correct set of properties.

In addition, we no longer need to build the SDK to extract the types
(that gets done in the script) or run api-extractor, so the scripts
get simpler and faster.  Also the TS compiler produces properties in
the order that we specified them in the file, which is nice.

The actual compiler API is particularly complicated and almost
entirely undocumented[1], which means this code is almost certainly
not the ideal way to extract some of this data, but it seems to work.
In particular, the compiler API operates in three different domains
(AST, symbol table, and type system), and you can get different
information from different domains, and possibly I could go between
the domains less if I knew what I was doing more.  But it should work,
and can be improved in the future.

[1] As an example, see
https://github.com/Microsoft/TypeScript/issues/15976#issuecomment-303168622,
which contains the sentence "I cannot coherently explain the naming
convention in a way that would let anyone intuit this behavior."

Test Plan:

npm run test

Compared all the REST API and JS API docs pages.  They all mostly
looked the same, and where they didn't look the same, they looked
better.

monorepo-commit: e313a4a69f976ed817e8b60fe8d7a18133b0b798