Action message ui3 component and usage in Message component
The new message component did not handle action messages correctly. This PR creates a new ActionMessage component that also uses vanilla css and replaces it within the message component!

Note: For context, this is not a primitive component. The styles that I have written in this PR work in the context of the cord-message, but this component will not work as a standalone. To be fair though, its only usage is inside the message component. I preferred to take the approach of removing a bunch of divs to make this component more lightweight than copying and pasting all the wrappers, which at the end of the day are unnecessary since there is only a single place usage of this component.

Test Plan:
Before:
<img width="608" alt="image" src="https://github.com/getcord/monorepo/assets/102963329/355d2ce4-4984-43af-addc-bedfbe52a792">

After:
<img width="605" alt="image" src="https://github.com/getcord/monorepo/assets/102963329/f2fa5c40-8884-44ed-acf8-65a84f115342">


Reviewers: AlbertQM, flooey, lazybean

Reviewed By: AlbertQM

Pull Request: https://github.com/getcord/monorepo/pull/5379

monorepo-commit: 5b310e51d15f0d7f4d1c09261ec63b059d12fbfb