ui3 TypingIndicator
ui3 TypingIndicator 

Summary:

CopyPaste and css convert
rm Text2, MessageBlockRow

A bit of annoying dance is needed because the facepile must take at least the width of the first column of message. But can grow out of it.
So instead of grid I'm using flex with min-width, and reuse the value of the message template through a css variables.


Test Plan:

see video

[TypingUser.mov](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/IoyKnJXsDnJREClofLDY/3321ac5c-00e2-4098-bac5-f56c87e1a922/TypingUser.mov)

monorepo-commit: 1237646cfef03f49616bd251fcb8fc245e671052